1 query for cookie_obj via WorkWithLink

+++ NW_Upvoter/db_tortories_orm/query/record.py
@@ -15,4 +15,7 @@ async def db_exist_record_link_account(link_obj, cookie_obj) -> (bool, WorkAccou
 @db_connection_required
 async def db_delete_record_work_account_with_link(obj_record: WorkAccountWithLink):
     """obj if needed to delete when exception exists"""
-    await obj_record.delete()
+    if obj_record is not None:
+        await obj_record.delete()
+    else:
+        print("obj_record is not an instance of WorkAccountWithLink")
